Reorganize events in Rust (#1378)
