+++ nautilus_core/model/src/events/position/mod.rs
@@ -13,6 +13,18 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-pub mod account;
-pub mod order;
-pub mod position;
+use crate::events::position::{
+    changed::PositionChanged, closed::PositionClosed, opened::PositionOpened,
+};
+
+pub mod changed;
+pub mod closed;
+pub mod opened;
+
+pub mod state;
+
+pub enum PositionEvent {
+    PositionOpened(PositionOpened),
+    PositionChanged(PositionChanged),
+    PositionClosed(PositionClosed),
+}
